+++ lib/experimental/editor-settings.php
@@ -38,6 +38,13 @@ function gutenberg_enable_experiments() {
 	if ( gutenberg_is_experiment_enabled( 'gutenberg-media-editor' ) ) {
 		wp_add_inline_script( 'wp-block-editor', 'window.__experimentalMediaEditor = true', 'before' );
 	}
+	if ( gutenberg_is_experiment_enabled( 'gutenberg-media-subject-crop' ) ) {
+		wp_add_inline_script(
+			'wp-block-editor',
+			'window.__experimentalMediaSubjectDetectionUrl = ' . wp_json_encode( gutenberg_url( 'build/media-detection/' ) ) . ';',
+			'before'
+		);
+	}
 	if ( gutenberg_is_experiment_enabled( 'gutenberg-dashboard-widgets' ) ) {
 		wp_add_inline_script( 'wp-block-editor', 'window.__experimentalDashboardWidgets = true', 'before' );
 	}